Add draft survival analysis plot
Need to confirm that it is handling the dates correctly with various
aggregation levels. I believe that the enddates are being calculated as
a day since the start of the simulation, but need to confirm this.

Also need to check why the starts of the survival plots are identical
between the detection (outbreak) and null curves, and then change around
1 year in. With a perfect test, only importations will be caught in the
earliest parts of the time series, which will produce identical test
positive time series (and backwards EWS), assuming the same importation
rate, but this changes with R0 (difference between null and detection
time series).